feat(upgrade-packages): refactoring
