+++ packages/core/index.ts
@@ -4,3 +4,4 @@ export * from "./util";
 export * from "./templates";
 export * from "./prompt";
 export * from "./packages/PackageManager";
+export * from "./update";
