--- packages/core/types/PackageDefinition.ts
@@ -1,4 +1,0 @@
-export interface PackageDefinition {
-	trial: string;
-	licensed: string;
-}
