+++ packages/core/update/index.ts
@@ -1,0 +1,2 @@
+export * from "./package-resolve";
+export * from "./Update";
