+++ packages/igx-templates/IgniteUIForAngularTemplate.ts
@@ -1,9 +1,8 @@
 import {
-	AddTemplateArgs, App,
-	ControlExtraConfiguration, FS_TOKEN, IFileSystem, Template, TemplateDependency, TypeScriptFileUpdate, Util
+	AddTemplateArgs, App, ControlExtraConfiguration, FS_TOKEN, IFileSystem, NPM_ANGULAR,
+	NPM_DOCK_MANAGER, Template, TemplateDependency, TypeScriptFileUpdate, Util, resolveIgxPackage
 } from "@igniteui/cli-core";
 import * as path from "path";
-import { NPM_DOCK_MANAGER, NPM_PACKAGE, resolveIgxPackage } from "./package-resolve";
 
 export class IgniteUIForAngularTemplate implements Template {
 	public components: string[];
@@ -159,7 +158,7 @@ export class IgniteUIForAngularTemplate implements Template {
 		config["description"] = this.description;
 		config["cliVersion"] = Util.version();
 		config["camelCaseName"] = Util.camelCase(name);
-		config["igxPackage"] = resolveIgxPackage(NPM_PACKAGE);
+		config["igxPackage"] = resolveIgxPackage(NPM_ANGULAR);
 		config["dockManagerPackage"] = resolveIgxPackage(NPM_DOCK_MANAGER);
 
 		/** 'nameMerged' is never used igx templates, removed */
