+++ packages/igx-templates/igx-ts-legacy/projects/_base/index.ts
@@ -1,6 +1,5 @@
-import { ControlExtraConfiguration, ProjectTemplate, Util } from "@igniteui/cli-core";
+import { ControlExtraConfiguration, ProjectTemplate, Util, updateWorkspace } from "@igniteui/cli-core";
 import * as path from "path";
-import { updateWorkspace } from "../../../Update";
 
 export class BaseIgxProject implements ProjectTemplate {
 	public id: string = "base";
