+++ packages/igx-templates/index.ts
@@ -1,7 +1,5 @@
 import * as standalone from "./igx-ts";
 import * as legacy from "./igx-ts-legacy";
 
-export * from "./package-resolve";
 export * from "./IgniteUIForAngularTemplate";
-export * from "./Update";
 export default [standalone, legacy];
